GCP/AIAgent_Blueprint/requirements.txt
# author: Olivier "Walgierd" Trela
# version 1.0
# Last edited: 22.12.2025 r.
# This file lists the Python packages required to run the AI Agent.

#requirements.txt

# --- Web Framework ---
Flask==3.0.3

# --- Google Cloud & AI Libraries ---
google-generativeai==0.7.2
google-api-python-client==2.160.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.18.0
google-cloud-aiplatform==1.71.1

# --- HTTP Request Libraries ---
requests==2.32.4

# --- Libraries for computation and background tasks ---
numpy==1.26.4
rq==1.16.2
